First bunch of clang tidy

+++ src/map/tile.cpp
@@ -675,7 +675,7 @@ Tile::SynchTileList Tile::GetTilesToSynch()
     TileItem_NonEmpty *t = static_cast<TileItem_NonEmpty*>(item[i]);
 
     if (t->NeedSynch()) {
-      SynchTileInfo info = { i, t->GetSynchsum() };
+      SynchTileInfo info = { static_cast<uint16_t>(i), t->GetSynchsum() };
       list.push_back(info);
     }
   }
